+++ tests/html/layout/gallery.php
@@ -6,7 +6,7 @@
     <div class="topic">
       <h2 class="main-heading"><span>layout: </span>gallery</h2>
       <div class="ovh">
-        <div class="gallery-number demo-container">
+        <div class="num demo-container">
           <ul data-margin>
             <li class="order autoheight"><div class="cell" data-content></div></li>
             <li class="order autoheight"><div class="cell" data-content></div></li>
@@ -18,7 +18,19 @@
             <li class="order autoheight"><div class="cell" data-content></div></li>
           </ul>
         </div>
-        <div class="gallery-child demo-container">
+        <div class="num-0px demo-container">
+          <ul data-margin>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+            <li class="order autoheight"><div class="cell" data-content></div></li>
+          </ul>
+        </div>
+        <div class="child demo-container">
           <div data-margin>
             <div class="order"><div class="cell" data-content></div></div>
             <div class="order"><div class="cell" data-content></div></div>
@@ -30,7 +42,7 @@
             <div class="order"><div class="cell" data-content></div></div>
           </div>
         </div>
-        <div class="gallery-direction demo-container">
+        <div class="direction demo-container">
           <ul data-margin>
             <li class="order"><div class="cell" data-content></div></li>
             <li class="order"><div class="cell" data-content></div></li>
@@ -42,7 +54,7 @@
             <li class="order"><div class="cell" data-content></div></li>
           </ul>
         </div>
-        <div class="gallery-center-align demo-container">
+        <div class="center-align demo-container">
           <ul data-margin>
             <li class="order"><div class="cell" data-content></div></li>
             <li class="order"><div class="cell" data-content></div></li>
@@ -54,7 +66,7 @@
             <li class="order"><div class="cell" data-content></div></li>
           </ul>
         </div>
-        <div class="gallery-percentage-gutter demo-container">
+        <div class="percentage-gutter demo-container">
           <ul data-margin>
             <li class="order"><div class="cell" data-content></div></li>
             <li class="order"><div class="cell" data-content></div></li>
@@ -66,7 +78,7 @@
             <li class="order"><div class="cell" data-content></div></li>
           </ul>
         </div>
-        <div class="gallery-rem-gutter demo-container">
+        <div class="rem-gutter demo-container">
           <ul data-margin>
             <li class="order"><div class="cell" data-content></div></li>
             <li class="order"><div class="cell" data-content></div></li>
@@ -78,7 +90,7 @@
             <li class="order"><div class="cell" data-content></div></li>
           </ul>
         </div>
-        <div class="gallery-map demo-container">
+        <div class="map demo-container">
           <ul data-margin>
             <li class="order"><div class="cell" data-content></div></li>
             <li class="order"><div class="cell" data-content></div></li>