+++ tests/html/layout/liquid-3.php
@@ -69,6 +69,13 @@
             <aside class="order"><div class="cell" data-content=""></div></aside>
           </div>
         </div>
+        <div class="liquid3-gutter-0px demo-container">
+          <div data-margin="">
+            <div class="order"><div class="cell" data-content=""></div></div>
+            <div class="order"><div class="cell" data-content=""></div></div>
+            <div class="order"><div class="cell" data-content=""></div></div>
+          </div>
+        </div>
         <div class="liquid3-gutter-sameunit demo-container">
           <div data-margin="">
             <div class="order"><div class="cell" data-content=""></div></div>